account_manager/requirements.txt
torch
django
djangorestframework-simplejwt
djangorestframework
django-cors-headers
Pillow
transformers